asjhbd

+++ 3D Template/Assets/scripts/Health.cs
@@ -6,7 +6,7 @@ using UnityEngine.UI;
 
 public class Health : MonoBehaviour
 {
-    public int health, maxHealth;
+    public float health, maxHealth;
     public static event Action OnPlayerDamaged;
     public static event Action OnPlayerDeath;
     public Image healthBar;
